Get averaged matches

Get averaged matches

Usage

get_averaged_matches(
  matched_data,
  treatment,
  outcome,
  covariates,
  match_id = "match_id",
  weight = "wt"
)

Arguments

matched_data

Filtered matched data set obtained from get_filtered_matches()

treatment

Character. Name of the treatment column in data. Currently assumes 1 for treatment and 0 for control.

outcome

Character. Name of the outcome column in data.

covariates

Character vector. Vector of covariate column names from data to use in match

match_id

Character. The column name in matched_data that contains the matching identifier. Default is "match_id".

weight

Character. The column name in matched_data that contains the weight. Default is "wt".

Value

A data frame of averaged matches

Examples

matches <- get_filtered_matches(
  example_trial_data,
  treatment = "Z",
  covariates = c("age", "stage4", "site", "ecog", "prevTrt", "diseaseFree")
)
#> ℹ You did not provide a value for `id`, so we created a unique id called `id`
#> ℹ You did not provide an indicator of categorical variables via `ind_categorical`, so we guessed that stage4, site, ecog, prevTrt, and diseaseFree are categorical.
#> ! Dropping pair 0_168 and 1_1 because site did not exactly match between groups.
#> ! Dropping pair 0_168 and 1_1 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_46 and 1_27 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_183 and 1_28 because site did not exactly match between groups.
#> ! Dropping pair 0_183 and 1_28 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_78 and 1_34 because site did not exactly match between groups.
#> ! Dropping pair 0_78 and 1_34 because diseaseFree did not exactly match between groups.
#> ! Dropping pair 0_11 and 1_4 because diseaseFree did not exactly match between groups.
#> ! Dropping pair 0_53 and 1_39 because site did not exactly match between groups.
#> ! Dropping pair 0_53 and 1_39 because diseaseFree did not exactly match between groups.
#> ! Dropping pair 0_132 and 1_40 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_25 and 1_42 because ecog did not exactly match between groups.
#> ! Dropping pair 0_25 and 1_42 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_2 and 1_44 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_17 and 1_45 because site did not exactly match between groups.
#> ! Dropping pair 0_17 and 1_45 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_76 and 1_50 because site did not exactly match between groups.
#> ! Dropping pair 0_76 and 1_50 because diseaseFree did not exactly match between groups.
#> ! Dropping pair 0_87 and 1_55 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_123 and 1_57 because diseaseFree did not exactly match between groups.
#> ! Dropping pair 0_137 and 1_58 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_26 and 1_6 because site did not exactly match between groups.
#> ! Dropping pair 0_186 and 1_59 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_35 and 1_61 because site did not exactly match between groups.
#> ! Dropping pair 0_35 and 1_61 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_124 and 1_66 because site did not exactly match between groups.
#> ! Dropping pair 0_130 and 1_72 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_162 and 1_73 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_5 and 1_74 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_161 and 1_8 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_63 and 1_80 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_144 and 1_85 because site did not exactly match between groups.
#> ! Dropping pair 0_64 and 1_86 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_172 and 1_90 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_115 and 1_10 because site did not exactly match between groups.
#> ! Dropping pair 0_131 and 1_103 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_23 and 1_104 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_160 and 1_112 because site did not exactly match between groups.
#> ! Dropping pair 0_36 and 1_12 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_98 and 1_126 because site did not exactly match between groups.
#> ! Dropping pair 0_98 and 1_126 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_95 and 1_127 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_199 and 1_128 because stage4 did not exactly match between groups.
#> ! Dropping pair 0_140 and 1_134 because site did not exactly match between groups.
#> ! Dropping pair 0_140 and 1_134 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_151 and 1_139 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_181 and 1_148 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_9 and 1_158 because site did not exactly match between groups.
#> ! Dropping pair 0_9 and 1_158 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_119 and 1_14 because site did not exactly match between groups.
#> ! Dropping pair 0_113 and 1_163 because site did not exactly match between groups.
#> ! Dropping pair 0_56 and 1_167 because site did not exactly match between groups.
#> ! Dropping pair 0_54 and 1_170 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_191 and 1_171 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_100 and 1_15 because site did not exactly match between groups.
#> ! Dropping pair 0_37 and 1_178 because stage4 did not exactly match between groups.
#> ! Dropping pair 0_7 and 1_182 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_21 and 1_188 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_150 and 1_189 because ecog did not exactly match between groups.
#> ! Dropping pair 0_193 and 1_194 because site did not exactly match between groups.
#> ! Dropping pair 0_193 and 1_194 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_107 and 1_197 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_114 and 1_198 because prevTrt did not exactly match between groups.
get_averaged_matches(
  matches,
  treatment = "Z",
  outcome = "Y",
  covariates = c("age", "stage4", "site", "ecog", "prevTrt", "diseaseFree"))
#>    match delta           wt      age stage4 site ecog prevTrt diseaseFree
#> 1   1.10     0 0.0316568356 59.41387      1    3    1       1           0
#> 2  1.100     0 0.0202603748 63.94452      1    1    0       2           1
#> 3   1.11     0 0.3877962364 61.42535      1    1    0       1           0
#> 4   1.12     0 0.1142811766 70.84145      1    1    1       1           0
#> 5   1.13     1 0.0383047711 62.97044      1    3    1       1           0
#> 6   1.16     1 0.3241659968 60.66532      1    3    1       1           0
#> 7   1.17     0 0.0028491152 69.07084      1    2    1       4           0
#> 8   1.19    -1 0.0079142089 62.15182      1    1    0       1           1
#> 9   1.25     0 0.0050650937 68.72140      1    1    0       1           0
#> 10  1.33     0 0.0383047711 71.23189      1    1    1       1           1
#> 11  1.34    -1 0.0028491152 71.85751      0    3    1       1           0
#> 12  1.35     1 0.0810414992 59.41464      1    3    1       1           1
#> 13  1.39     0 0.0620473978 59.57441      1    1    1       1           1
#> 14  1.40    -1 0.2140002088 66.99852      1    3    1       1           0
#> 15  1.41     0 0.0810414992 63.69758      1    1    1       1           1
#> 16  1.43     0 0.0079142089 69.51796      1    1    1       1           1
#> 17  1.44     0 0.6698586418 68.14865      1    3    0       3           0
#> 18  1.45     0 1.0649359504 58.58658      1    2    1       1           1
#> 19  1.48    -1 0.0383047711 72.27510      1    1    1       1           0
#> 20  1.50     0 0.3877962364 60.41166      1    1    1       2           1
#> 21  1.53    -1 0.0455858433 72.48265      1    1    1       2           0
#> 22  1.54     0 0.7600806234 62.75114      1    1    0       3           0
#> 23  1.55    -1 0.6698586418 60.41382      1    1    0       1           0
#> 24  1.56     1 0.0028491152 61.70981      1    1    1       1           0
#> 25  1.57     0 0.1978552227 63.89357      1    1    1       1           0
#> 26  1.59     1 0.0383047711 71.70576      1    2    1       1           0
#> 27  1.60     0 0.0202603748 67.82238      1    2    1       1           0
#> 28  1.61    -1 1.2168887614 69.84563      1    1    0       2           0
#> 29  1.62    -1 0.0155118495 58.14970      1    1    1       1           1
#> 30  1.66     0 0.1025681474 66.64420      1    1    1       2           0
#> 31  1.68     1 0.0050650937 69.71459      1    3    0       1           0
#> 32  1.69     1 0.2481895913 64.34767      1    1    0       1           1
#> 33   1.7     0 0.1025681474 64.65735      1    1    1       2           0
#> 34  1.71    -1 0.0079142089 65.54090      1    2    1       1           0
#> 35  1.72     0 1.1779508536 65.22624      1    3    0       1           0
#> 36  1.73     0 0.0079142089 61.09855      1    1    1       1           0
#> 37  1.75     0 0.0113964608 62.73999      1    3    1       1           0
#> 38  1.76    -1 0.2140002088 62.27693      1    3    0       2           0
#> 39  1.77    -1 0.5584265804 68.82323      1    1    0       4           1
#> 40  1.78     0 0.4571247064 61.04249      1    1    0       1           0
#> 41  1.81     1 0.0535000522 72.89905      1    1    1       1           1
#> 42  1.82     0 0.5584265804 58.92139      1    3    1       2           0
#> 43  1.83     0 0.7600806234 62.18283      1    3    1       1           1
#> 44  1.85    -1 0.5584265804 71.03553      1    1    1       4           0
#> 45  1.88     0 0.1978552227 72.14927      1    1    0       1           1
#> 46  1.89     1 0.1142811766 68.49060      1    3    0       1           0
#> 47  1.92    -1 0.0003165684 65.72637      1    2    1       4           1
#> 48  1.95     0 0.2481895913 67.40937      1    1    1       2           0
#> 49  1.97     1 0.0535000522 65.11422      1    1    1       1           0