Get filtered matches

Get filtered matches

Usage

get_filtered_matches(
  data,
  treatment,
  covariates,
  id = NULL,
  weight = "wt",
  ind_categorical = NULL
)

Arguments

data

Data frame.

treatment

Character. Name of the treatment column in data. Currently assumes 1 for treatment and 0 for control.

covariates

Character vector. Vector of covariate column names from data to use in match

id

Character. Name of the id column in data

weight

Character. The name that should be given to the variable containing the weights in the output data frame. Default is "wt".

ind_categorical

Logical vector. (Optional) Indicator of which . covariates are categorical.

Value

A data frame of filtered matches

Examples

get_filtered_matches(
  example_trial_data,
  treatment = "Z",
  covariates = c("age", "stage4", "site", "ecog", "prevTrt", "diseaseFree")
)
#> ℹ You did not provide a value for `id`, so we created a unique id called `id`
#> ℹ You did not provide an indicator of categorical variables via `ind_categorical`, so we guessed that stage4, site, ecog, prevTrt, and diseaseFree are categorical.
#> ! Dropping pair 0_168 and 1_1 because site did not exactly match between groups.
#> ! Dropping pair 0_168 and 1_1 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_46 and 1_27 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_183 and 1_28 because site did not exactly match between groups.
#> ! Dropping pair 0_183 and 1_28 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_78 and 1_34 because site did not exactly match between groups.
#> ! Dropping pair 0_78 and 1_34 because diseaseFree did not exactly match between groups.
#> ! Dropping pair 0_11 and 1_4 because diseaseFree did not exactly match between groups.
#> ! Dropping pair 0_53 and 1_39 because site did not exactly match between groups.
#> ! Dropping pair 0_53 and 1_39 because diseaseFree did not exactly match between groups.
#> ! Dropping pair 0_132 and 1_40 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_25 and 1_42 because ecog did not exactly match between groups.
#> ! Dropping pair 0_25 and 1_42 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_2 and 1_44 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_17 and 1_45 because site did not exactly match between groups.
#> ! Dropping pair 0_17 and 1_45 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_76 and 1_50 because site did not exactly match between groups.
#> ! Dropping pair 0_76 and 1_50 because diseaseFree did not exactly match between groups.
#> ! Dropping pair 0_87 and 1_55 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_123 and 1_57 because diseaseFree did not exactly match between groups.
#> ! Dropping pair 0_137 and 1_58 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_26 and 1_6 because site did not exactly match between groups.
#> ! Dropping pair 0_186 and 1_59 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_35 and 1_61 because site did not exactly match between groups.
#> ! Dropping pair 0_35 and 1_61 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_124 and 1_66 because site did not exactly match between groups.
#> ! Dropping pair 0_130 and 1_72 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_162 and 1_73 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_5 and 1_74 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_161 and 1_8 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_63 and 1_80 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_144 and 1_85 because site did not exactly match between groups.
#> ! Dropping pair 0_64 and 1_86 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_172 and 1_90 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_115 and 1_10 because site did not exactly match between groups.
#> ! Dropping pair 0_131 and 1_103 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_23 and 1_104 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_160 and 1_112 because site did not exactly match between groups.
#> ! Dropping pair 0_36 and 1_12 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_98 and 1_126 because site did not exactly match between groups.
#> ! Dropping pair 0_98 and 1_126 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_95 and 1_127 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_199 and 1_128 because stage4 did not exactly match between groups.
#> ! Dropping pair 0_140 and 1_134 because site did not exactly match between groups.
#> ! Dropping pair 0_140 and 1_134 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_151 and 1_139 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_181 and 1_148 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_9 and 1_158 because site did not exactly match between groups.
#> ! Dropping pair 0_9 and 1_158 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_119 and 1_14 because site did not exactly match between groups.
#> ! Dropping pair 0_113 and 1_163 because site did not exactly match between groups.
#> ! Dropping pair 0_56 and 1_167 because site did not exactly match between groups.
#> ! Dropping pair 0_54 and 1_170 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_191 and 1_171 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_100 and 1_15 because site did not exactly match between groups.
#> ! Dropping pair 0_37 and 1_178 because stage4 did not exactly match between groups.
#> ! Dropping pair 0_7 and 1_182 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_21 and 1_188 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_150 and 1_189 because ecog did not exactly match between groups.
#> ! Dropping pair 0_193 and 1_194 because site did not exactly match between groups.
#> ! Dropping pair 0_193 and 1_194 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_107 and 1_197 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_114 and 1_198 because prevTrt did not exactly match between groups.
#>          age stage4 site prevTrt ecog diseaseFree delta Y1 Y0 UNIQID Z Y  id
#> 3   58.64473      1    3       1    1           0     1  1  0     A3 0 0   3
#> 16  60.18301      1    3       1    1           0    -1  0  1    A16 1 0  16
#> 33  63.77872      1    1       2    0           1     0  1  1    A33 0 1  33
#> 200 64.11032      1    1       2    0           1     0  1  1   A200 1 1 200
#> 190 62.77406      1    1       1    0           0     0  1  1   A190 0 1 190
#> 18  60.07663      1    1       1    0           0     0  1  1    A18 1 1  18
#> 70  72.19511      1    1       1    1           0    -1  0  1    A70 0 1  70
#> 19  69.48780      1    1       1    1           0     1  1  0    A19 1 1  19
#> 145 62.65232      1    3       1    1           0     1  1  0   A145 0 0 145
#> 22  63.28855      1    3       1    1           0     1  1  0    A22 1 1  22
#> 41  59.08793      1    3       1    1           0     1  1  0    A41 0 0  41
#> 30  62.24271      1    3       1    1           0     0  1  1    A30 1 1  30
#> 96  69.19607      1    2       4    1           0    -1  0  1    A96 0 1  96
#> 32  68.94562      1    2       4    1           0     1  1  0    A32 1 1  32
#> 97  61.94202      1    1       1    0           1    -1  0  1    A97 0 1  97
#> 38  62.36162      1    1       1    0           1    -1  0  1    A38 1 0  38
#> 102 68.99171      1    1       1    0           0     0  1  1   A102 0 1 102
#> 48  68.45108      1    1       1    0           0     1  1  0    A48 1 1  48
#> 52  72.01675      1    1       1    1           1    -1  0  1    A52 0 1  52
#> 67  70.44703      1    1       1    1           1     1  1  0    A67 1 1  67
#> 142 71.60989      0    3       1    1           0    -1  0  1   A142 0 1 142
#> 69  72.10513      0    3       1    1           0    -1  0  1    A69 1 0  69
#> 93  60.30583      1    3       1    1           1     1  1  0    A93 0 0  93
#> 71  58.52344      1    3       1    1           1     0  1  1    A71 1 1  71
#> 65  58.68590      1    1       1    1           1    -1  0  1    A65 0 1  65
#> 75  60.46292      1    1       1    1           1     1  1  0    A75 1 1  75
#> 180 67.73535      1    3       1    1           0    -1  0  1   A180 0 1 180
#> 77  66.26170      1    3       1    1           0    -1  0  1    A77 1 0  77
#> 47  64.12984      1    1       1    1           1    -1  0  1    A47 0 1  47
#> 79  63.26532      1    1       1    1           1     1  1  0    A79 1 1  79
#> 149 69.26814      1    1       1    1           1    -1  0  1   A149 0 1 149
#> 82  69.76779      1    1       1    1           1     0  1  1    A82 1 1  82
#> 147 69.99641      1    3       3    0           0     1  1  0   A147 0 0 147
#> 83  66.30089      1    3       3    0           0    -1  0  1    A83 1 0  83
#> 99  62.42383      1    2       1    1           1     0  1  1    A99 0 1  99
#> 84  54.74933      1    2       1    1           1     1  1  0    A84 1 1  84
#> 155 73.31435      1    1       1    1           0     0  1  1   A155 0 1 155
#> 88  71.23585      1    1       1    1           0    -1  0  1    A88 1 0  88
#> 185 58.62592      1    1       2    1           1    -1  0  1   A185 0 1 185
#> 91  62.19740      1    1       2    1           1     1  1  0    A91 1 1  91
#> 129 71.49756      1    1       2    1           0     0  1  1   A129 0 1 129
#> 105 73.46774      1    1       2    1           0    -1  0  1   A105 1 0 105
#> 157 64.34463      1    1       3    0           0    -1  0  1   A157 0 1 157
#> 106 61.15765      1    1       3    0           0     0  1  1   A106 1 1 106
#> 111 62.43064      1    1       1    0           0    -1  0  1   A111 0 1 111
#> 108 58.39700      1    1       1    0           0    -1  0  1   A108 1 0 108
#> 173 61.79873      1    1       1    1           0     1  1  0   A173 0 0 173
#> 109 61.62089      1    1       1    1           0     1  1  0   A109 1 1 109
#> 138 64.62216      1    1       1    1           0    -1  0  1   A138 0 1 138
#> 110 63.16497      1    1       1    1           0     0  1  1   A110 1 1 110
#> 117 72.49181      1    2       1    1           0     1  1  0   A117 0 0 117
#> 116 70.91970      1    2       1    1           0     1  1  0   A116 1 1 116
#> 159 68.26181      1    2       1    1           0    -1  0  1   A159 0 1 159
#> 121 67.38295      1    2       1    1           0     1  1  0   A121 1 1 121
#> 94  73.17825      1    1       2    0           0     0  1  1    A94 0 1  94
#> 122 66.51302      1    1       2    0           0    -1  0  1   A122 1 0 122
#> 51  58.50165      1    1       1    1           1     0  1  1    A51 0 1  51
#> 125 57.79775      1    1       1    1           1    -1  0  1   A125 1 0 125
#> 49  67.06692      1    1       2    1           0    -1  0  1    A49 0 1  49
#> 133 66.22148      1    1       2    1           0     0  1  1   A133 1 1 133
#> 175 69.99359      1    3       1    0           0     1  1  0   A175 0 0 175
#> 135 69.43560      1    3       1    0           0     0  1  1   A135 1 1 135
#> 118 63.55448      1    1       1    0           1     1  1  0   A118 0 0 118
#> 136 65.14086      1    1       1    0           1     0  1  1   A136 1 1 136
#> 179 65.19831      1    1       2    1           0    -1  0  1   A179 0 1 179
#> 13  64.11638      1    1       2    1           0     1  1  0    A13 1 1  13
#> 187 65.63329      1    2       1    1           0     0  1  1   A187 0 1 187
#> 141 65.44852      1    2       1    1           0    -1  0  1   A141 1 0 141
#> 81  66.99084      1    3       1    0           0     0  1  1    A81 0 1  81
#> 143 63.46163      1    3       1    0           0     1  1  0   A143 1 1 143
#> 24  61.52631      1    1       1    1           0    -1  0  1    A24 0 1  24
#> 146 60.67078      1    1       1    1           0     0  1  1   A146 1 1 146
#> 92  62.45935      1    3       1    1           0     1  1  0    A92 0 0  92
#> 152 63.02063      1    3       1    1           0    -1  0  1   A152 1 0 152
#> 68  61.50261      1    3       2    0           0    -1  0  1    A68 0 1  68
#> 153 63.05126      1    3       2    0           0    -1  0  1   A153 1 0 153
#> 60  70.72859      1    1       4    0           1    -1  0  1    A60 0 1  60
#> 154 66.91787      1    1       4    0           1    -1  0  1   A154 1 0 154
#> 177 62.72568      1    1       1    0           0    -1  0  1   A177 0 1 177
#> 156 59.35930      1    1       1    0           0     0  1  1   A156 1 1 156
#> 29  74.00262      1    1       1    1           1     1  1  0    A29 0 0  29
#> 164 71.79548      1    1       1    1           1     0  1  1   A164 1 1 164
#> 89  56.14904      1    3       2    1           0    -1  0  1    A89 0 1  89
#> 165 61.69375      1    3       2    1           0     0  1  1   A165 1 1 165
#> 43  63.93857      1    3       1    1           1     0  1  1    A43 0 1  43
#> 166 60.42709      1    3       1    1           1     0  1  1   A166 1 1 166
#> 195 73.72058      1    1       4    1           0    -1  0  1   A195 0 1 195
#> 169 68.35049      1    1       4    1           0    -1  0  1   A169 1 0 169
#> 62  70.28421      1    1       1    0           1    -1  0  1    A62 0 1  62
#> 174 74.01433      1    1       1    0           1     0  1  1   A174 1 1 174
#> 120 67.59256      1    3       1    0           0     1  1  0   A120 0 0 120
#> 176 69.38864      1    3       1    0           0     1  1  0   A176 1 1 176
#> 20  65.64631      1    2       4    1           1     0  1  1    A20 0 1  20
#> 184 65.80643      1    2       4    1           1    -1  0  1   A184 1 0 184
#> 101 68.41959      1    1       2    1           0     0  1  1   A101 0 1 101
#> 192 66.39915      1    1       2    1           0     1  1  0   A192 1 1 192
#> 31  65.53529      1    1       1    1           0     1  1  0    A31 0 0  31
#> 196 64.69315      1    1       1    1           0     0  1  1   A196 1 1 196
#>     match_id           wt
#> 3       1.10 0.0316568356
#> 16      1.10 0.0316568356
#> 33     1.100 0.0202603748
#> 200    1.100 0.0202603748
#> 190     1.11 0.3877962364
#> 18      1.11 0.3877962364
#> 70      1.12 0.1142811766
#> 19      1.12 0.1142811766
#> 145     1.13 0.0383047711
#> 22      1.13 0.0383047711
#> 41      1.16 0.3241659968
#> 30      1.16 0.3241659968
#> 96      1.17 0.0028491152
#> 32      1.17 0.0028491152
#> 97      1.19 0.0079142089
#> 38      1.19 0.0079142089
#> 102     1.25 0.0050650937
#> 48      1.25 0.0050650937
#> 52      1.33 0.0383047711
#> 67      1.33 0.0383047711
#> 142     1.34 0.0028491152
#> 69      1.34 0.0028491152
#> 93      1.35 0.0810414992
#> 71      1.35 0.0810414992
#> 65      1.39 0.0620473978
#> 75      1.39 0.0620473978
#> 180     1.40 0.2140002088
#> 77      1.40 0.2140002088
#> 47      1.41 0.0810414992
#> 79      1.41 0.0810414992
#> 149     1.43 0.0079142089
#> 82      1.43 0.0079142089
#> 147     1.44 0.6698586418
#> 83      1.44 0.6698586418
#> 99      1.45 1.0649359504
#> 84      1.45 1.0649359504
#> 155     1.48 0.0383047711
#> 88      1.48 0.0383047711
#> 185     1.50 0.3877962364
#> 91      1.50 0.3877962364
#> 129     1.53 0.0455858433
#> 105     1.53 0.0455858433
#> 157     1.54 0.7600806234
#> 106     1.54 0.7600806234
#> 111     1.55 0.6698586418
#> 108     1.55 0.6698586418
#> 173     1.56 0.0028491152
#> 109     1.56 0.0028491152
#> 138     1.57 0.1978552227
#> 110     1.57 0.1978552227
#> 117     1.59 0.0383047711
#> 116     1.59 0.0383047711
#> 159     1.60 0.0202603748
#> 121     1.60 0.0202603748
#> 94      1.61 1.2168887614
#> 122     1.61 1.2168887614
#> 51      1.62 0.0155118495
#> 125     1.62 0.0155118495
#> 49      1.66 0.1025681474
#> 133     1.66 0.1025681474
#> 175     1.68 0.0050650937
#> 135     1.68 0.0050650937
#> 118     1.69 0.2481895913
#> 136     1.69 0.2481895913
#> 179      1.7 0.1025681474
#> 13       1.7 0.1025681474
#> 187     1.71 0.0079142089
#> 141     1.71 0.0079142089
#> 81      1.72 1.1779508536
#> 143     1.72 1.1779508536
#> 24      1.73 0.0079142089
#> 146     1.73 0.0079142089
#> 92      1.75 0.0113964608
#> 152     1.75 0.0113964608
#> 68      1.76 0.2140002088
#> 153     1.76 0.2140002088
#> 60      1.77 0.5584265804
#> 154     1.77 0.5584265804
#> 177     1.78 0.4571247064
#> 156     1.78 0.4571247064
#> 29      1.81 0.0535000522
#> 164     1.81 0.0535000522
#> 89      1.82 0.5584265804
#> 165     1.82 0.5584265804
#> 43      1.83 0.7600806234
#> 166     1.83 0.7600806234
#> 195     1.85 0.5584265804
#> 169     1.85 0.5584265804
#> 62      1.88 0.1978552227
#> 174     1.88 0.1978552227
#> 120     1.89 0.1142811766
#> 176     1.89 0.1142811766
#> 20      1.92 0.0003165684
#> 184     1.92 0.0003165684
#> 101     1.95 0.2481895913
#> 192     1.95 0.2481895913
#> 31      1.97 0.0535000522
#> 196     1.97 0.0535000522