Calculate Effect Statistics

Calculate Effect Statistics

Usage

get_mtree_stats(data, treatment, outcome, covariates, tree, cens_var = NULL)

Arguments

data

Data frame.

treatment

Character. Name of the treatment column in data

outcome

Character. Name of the outcome column in data

covariates

Character vector. Vector of covariate column names from data to use in match

tree

Decision tree obtained from get_mtree().

cens_var

Character. Name of a censoring column in data. If none exists, leave NULL. Default: NULL

Value

Data frame with effect estimates

Examples

covariates <- c("age", "stage4", "site", "ecog", "prevTrt", "diseaseFree")
matches <- get_filtered_matches(
  example_trial_data,
  treatment = "Z",
  covariates = covariates
)
#> ℹ You did not provide a value for `id`, so we created a unique id called `id`
#> ℹ You did not provide an indicator of categorical variables via `ind_categorical`, so we guessed that stage4, site, ecog, prevTrt, and diseaseFree are categorical.
#> ! Dropping pair 0_168 and 1_1 because site did not exactly match between groups.
#> ! Dropping pair 0_168 and 1_1 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_46 and 1_27 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_183 and 1_28 because site did not exactly match between groups.
#> ! Dropping pair 0_183 and 1_28 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_78 and 1_34 because site did not exactly match between groups.
#> ! Dropping pair 0_78 and 1_34 because diseaseFree did not exactly match between groups.
#> ! Dropping pair 0_11 and 1_4 because diseaseFree did not exactly match between groups.
#> ! Dropping pair 0_53 and 1_39 because site did not exactly match between groups.
#> ! Dropping pair 0_53 and 1_39 because diseaseFree did not exactly match between groups.
#> ! Dropping pair 0_132 and 1_40 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_25 and 1_42 because ecog did not exactly match between groups.
#> ! Dropping pair 0_25 and 1_42 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_2 and 1_44 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_17 and 1_45 because site did not exactly match between groups.
#> ! Dropping pair 0_17 and 1_45 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_76 and 1_50 because site did not exactly match between groups.
#> ! Dropping pair 0_76 and 1_50 because diseaseFree did not exactly match between groups.
#> ! Dropping pair 0_87 and 1_55 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_123 and 1_57 because diseaseFree did not exactly match between groups.
#> ! Dropping pair 0_137 and 1_58 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_26 and 1_6 because site did not exactly match between groups.
#> ! Dropping pair 0_186 and 1_59 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_35 and 1_61 because site did not exactly match between groups.
#> ! Dropping pair 0_35 and 1_61 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_124 and 1_66 because site did not exactly match between groups.
#> ! Dropping pair 0_130 and 1_72 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_162 and 1_73 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_5 and 1_74 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_161 and 1_8 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_63 and 1_80 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_144 and 1_85 because site did not exactly match between groups.
#> ! Dropping pair 0_64 and 1_86 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_172 and 1_90 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_115 and 1_10 because site did not exactly match between groups.
#> ! Dropping pair 0_131 and 1_103 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_23 and 1_104 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_160 and 1_112 because site did not exactly match between groups.
#> ! Dropping pair 0_36 and 1_12 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_98 and 1_126 because site did not exactly match between groups.
#> ! Dropping pair 0_98 and 1_126 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_95 and 1_127 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_199 and 1_128 because stage4 did not exactly match between groups.
#> ! Dropping pair 0_140 and 1_134 because site did not exactly match between groups.
#> ! Dropping pair 0_140 and 1_134 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_151 and 1_139 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_181 and 1_148 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_9 and 1_158 because site did not exactly match between groups.
#> ! Dropping pair 0_9 and 1_158 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_119 and 1_14 because site did not exactly match between groups.
#> ! Dropping pair 0_113 and 1_163 because site did not exactly match between groups.
#> ! Dropping pair 0_56 and 1_167 because site did not exactly match between groups.
#> ! Dropping pair 0_54 and 1_170 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_191 and 1_171 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_100 and 1_15 because site did not exactly match between groups.
#> ! Dropping pair 0_37 and 1_178 because stage4 did not exactly match between groups.
#> ! Dropping pair 0_7 and 1_182 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_21 and 1_188 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_150 and 1_189 because ecog did not exactly match between groups.
#> ! Dropping pair 0_193 and 1_194 because site did not exactly match between groups.
#> ! Dropping pair 0_193 and 1_194 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_107 and 1_197 because prevTrt did not exactly match between groups.
#> ! Dropping pair 0_114 and 1_198 because prevTrt did not exactly match between groups.
m <- get_averaged_matches(
  matches,
  treatment = "Z",
  outcome = "Y",
  covariates = covariates
  )
tree <- get_mtree(m, covariates)
get_mtree_stats(example_trial_data,
  treatment = "Z",
  outcome = "Y",
  covariates = covariates,
  tree = tree)
#> $est
#> data$predNode: 1
#>   x1  n1   p1 x0  n0   p0  diff lower  upper
#> 1 59 100 0.59 76 100 0.76 -0.17 -0.27 -0.045
#>